Remove d2l-filter-dropdown-page and create d2l-filter-dropdown-category, which does not store the data about the options
